Fix UID and GID for macOS Docker
* Replace UID and GID variables to MYUID and MYGID

* Use default GID of 1000 on macOS (Darwin)

The crossplatform problem has two parts:

1. On mac, the UID and GID environment variables are read-only,
as they are already produced by the system. Therefore the
local variable needs to be renamed (for example MYUID)

2. If the mac user is part of the "staff" user group (root),
the GID will be 20. In ubuntu this is a reserved group ("dialout"),
so the docker build step addgroup --gid ${GID} will fail.

Conversely though, mac also resolves the whole permissions thing
nicely in a different way:

> Inside the container, the osxfs driver pretends that whatever
uid/gid is running in the container is also the uid/gid that
owns the mounted files.
> If you chown files (in the container) to something else,
no chown is performed on the real files; this owner information
is stored in an extended file attribute instead,
and that value is used by the container (and ignored by macOS).
> The real files are owned by whoever owns them in macOS,
outside the container. Access control is determined using
those real file ownerships, and the uid/gid of the user running
the Docker applications (which is probably your login user)