Fix bug in DS twist conversion (#66)
The return of the compute_command function in Ring
and Circular had `CartesianTwist(center) * twist` to
return the local twist back to the base frame.

But, this makes no sense! The construction of
a twist from pose actually does unit differentiation,
so it ends up returning a constant bias twist
that increases the further the center is from
Identity.

Really I intended to convert it to a State, not a pose,
so that the center pose transforms the local twist.

The unit tests didn't catch it because the center
is set to Identity. This makes checking the expected
outputs at certain points easier at least.

I have added a test to check convergence on radius
for a random center pose for Ring and Circular.