Fix rotation offset transformation order in Ring DS
* Get local pose in rotation offset frame, instead of inverse

The rotation offset in the Ring DS is a bit weird to understand,
but it's definitely necessary for configuring rotating frames around
the ring center.

For a given Pose in Base frame, you can get the Local Pose in Base
frame through this series of transformations:
Center in Base frame
Rotation offset in Center frame
Local Pose in Rotation offset frame

I added more test cases to cover this behaviour.