Fix Ring DS rotation offset and boundary condition (#90)
* Fix rotation offset transformation order in Ring DS

* Get local pose in rotation offset frame, instead of inverse

The rotation offset in the Ring DS is a bit weird to understand,
but it's definitely necessary for configuring rotating frames around
the ring center.

For a given Pose in Base frame, you can get the Local Pose in Base
frame through this series of transformations:
Center in Base frame
Rotation offset in Center frame
Local Pose in Rotation offset frame

I added more test cases to cover this behaviour.

* Fix angular displacement volatility at boundary

As the ring produces linear velocity around the center,
this also induces an instantaneous angular velocity around the
local Z axis. This is calculated with a dot product between
the current position on the circular plane and the "future"
position if the position was integrated with the velocity.

Unfortunately, if the speed of the ring DS is very high,
a linear velocity pointing in the direction of the circle
center would appear to end up on the other side of the circle,
resulting in an apparent angular displacement of ~pi.

The main side-effect of this is jumpy angular twist output
at the boundary of radius + width when desired speed is large.

By clamping the local linear velocity based on the position in
the circle, it prevents this projection from ever getting close
to or beyond the circle center where the angular displacement
is much more volatile, resulting in smoother transitions
at the ring boundary.