Fix CartesianPose constructor and revise run script (#157)
* Add custom lambda bindings for CartesianPose constructors
involving the orientation. Because Eigen::Quaterniond is not
implicitly convertible by PyBind11, an Eigen::Vector4d is
used as a placeholder instead.

* Add a cartesian_pose_test file to test constructors

* Improve run script for testing bindings by adding a
branch argument to the clone stage of control libraries
in the Dockerfile. The branch is the current local one
by default, else is specified by the first input argument.

* Mount the source volume when running so that it's easier
to test changes and rebuild quickly

* Update python README

* Update CHANGELOG