Make clproto shared again, remove header implementations
* Mark clproto as a shared library

* Set Position Independent Code to ON for
clproto_bindings only

* Remove NotImplementedException and default template
implementation for encode / decode, since this
causes the dynamic linker to replace *ALL* calls
to encode / decode with the "NotImplemented" definition,
even if it is implemented in clproto. By removing
this, the compile (link) will fail if an invalid
object is encoded / decoded, instead of raising
a run-time exception, but at least valid objects
will work. Note that the above did not apply
to the static version of clproto, but the
static one had other downstream link issues.

* Remove associated NotImplemented test and leave
a descriptive comment field instead