Revert templated en/decoders
* Remove templated encoder and
decoder helper methods (of the
form `MsgT encode(ObjT)` and
`ObjT decode(MsgT)` in favor
of explicilty declared and
overloaded methods. This is
because in release build, the
templated function definitions
get lost without additional
forward declaration / special
instantiation. This would make
template usage far more verbose
than necessary, and so it is easier
just to declare the overloaded
methods in the respective helper
header files.

* CI workflow to test in release build