Selective component dependencies
* In cmake package configuration, check for robot model dependencies
only if no components are listed in the find_package step. Otherwise,
if components are listed and include robot model or controller,
find the corresponding dependencies

* Update clproto cmakelists to specifically subselect the
state_representation component, so that the find_dependencies
for pinocchio are not triggered. This would otherwise cause a
failure because /opt/openrobots is not on the prefix path.

* Update README with component selection info