Selective component dependencies (#261)
* Selective component dependencies

* In cmake package configuration, check for robot model dependencies
only if no components are listed in the find_package step. Otherwise,
if components are listed and include robot model or controller,
find the corresponding dependencies

* Update clproto cmakelists to specifically subselect the
state_representation component, so that the find_dependencies
for pinocchio are not triggered. This would otherwise cause a
failure because /opt/openrobots is not on the prefix path.

* Update README with component selection info

* Fix update version script to find and replace the project version
correctly using the project name with underscore `control_libraries`

* Set component found variable

* If a valid component is listed, set a specific named variable
to indicate that component was found, so that the macro
check_required_components will succeed

* Changelog

* 5.0.3 -> 5.0.4